requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
Babel==2.9.1
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-modeltranslation==0.17.2
django-tinymce==3.3.0
idna==2.10
Markdown==3.3.4
multidict==5.1.0
Pillow==8.2.0
pyTelegramBotAPI==3.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
